requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.7.0
PyExifTool==0.1
Werkzeug==0.9.6
